Convert `ExtendedModelTemplate.Builder` to being lambda based
Split out all inner classes of `ExtendedModelTemplate` into their own classes
