+++ src/main/java/net/neoforged/neoforge/client/model/generators/template/package-info.java
@@ -6,7 +6,7 @@
 @FieldsAreNonnullByDefault
 @MethodsReturnNonnullByDefault
 @ParametersAreNonnullByDefault
-package net.neoforged.neoforge.client.model.generators;
+package net.neoforged.neoforge.client.model.generators.template;
 
 import javax.annotation.ParametersAreNonnullByDefault;
 import net.minecraft.FieldsAreNonnullByDefault;
